move more stuff into the server-shared package
