+++ apps/passport-server/src/apis/honeycombAPI.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line @typescript-eslint/triple-slash-reference
 /// <reference path="../util/declarations/libhoney.d.ts" />
-import { requireEnv } from "@pcd/util";
+import { requireEnv } from "@pcd/server-shared";
 import Libhoney from "libhoney";
 import { ApplicationContext } from "../types";
 import { logger } from "../util/logger";
