+++ apps/passport-server/src/services/startRollbarService.ts
@@ -1,5 +1,4 @@
-import { RollbarService } from "@pcd/server-shared";
-import { requireEnv } from "@pcd/util";
+import { RollbarService, requireEnv } from "@pcd/server-shared";
 import { ApplicationContext } from "../types";
 import { logger } from "../util/logger";
